+++ components/Model/DTP/CardContent/components/Participants/Participant/Participant.module.css
