+++ components/Model/DTP/CardContent/components/Participants/Participants.types.ts
@@ -2,9 +2,6 @@ import { DTPObjectAttributes } from 'common/data/dtp/dtp';
 
 export type ParticipantsProps = {
     participants: DTPObjectAttributes['participants'];
-    participantsCount: number;
-    injuredCount: number;
-    deadCount: number;
     vehicles?: DTPObjectAttributes['vehicles'];
 };
