+++ components/Model/DTP/MapData/MapData.types.ts
@@ -1,0 +1,6 @@
+import { LatLngExpression } from 'leaflet';
+
+export type DTPMapDataProps = {
+    id: string;
+    coords: LatLngExpression;
+};
