+++ components/Model/Houses/MapData/MapData.types.ts
@@ -1,0 +1,4 @@
+export type HousesMapDataProps = {
+    borders: [number, number][];
+    id: string;
+};
