+++ components/Model/OKN/CardContent/CardContent.module.css
@@ -1,5 +1,4 @@
 .popup {
-    font-family: 'Iset Sans', sans-serif;
     border-top-left-radius: inherit;
     border-top-right-radius: inherit;
 }
