+++ components/Model/OKN/MapData/MapData.types.ts
@@ -1,0 +1,6 @@
+import { LatLngExpression } from 'leaflet';
+
+export type OKNMapDataProps = {
+    id: string;
+    coords: LatLngExpression[] | LatLngExpression;
+};
